fix: move warehouse size enum
